# apps/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/user/me": {
      "get": {
        "operationId": "UsersController_getMe",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": ["Users"]
      }
    },
    "/api/user/profile": {
      "patch": {
        "operationId": "UsersController_updateProfile",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Function"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": ["Users"]
      }
    },
    "/api/user/avatar": {
      "post": {
        "operationId": "UsersController_uploadAvatar",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": ["Users"]
      }
    },
    "/api/user/preferences": {
      "patch": {
        "operationId": "UsersController_updatePreferences",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Function"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": ["Users"]
      }
    },
    "/api/user/reader-preferences": {
      "patch": {
        "operationId": "UsersController_updateReaderPreferences",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Function"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": ["Users"]
      }
    },
    "/api/user/notifications": {
      "patch": {
        "operationId": "UsersController_updateNotifications",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Function"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": ["Users"]
      }
    },
    "/api/user/privacy": {
      "patch": {
        "operationId": "UsersController_updatePrivacy",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Function"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": ["Users"]
      }
    },
    "/api/user/mood": {
      "patch": {
        "operationId": "UsersController_updateMood",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Function"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": ["Users"]
      }
    },
    "/api/user/email-change-request": {
      "post": {
        "operationId": "UsersController_requestEmailChange",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Function"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": ["Users"]
      }
    },
    "/api/user/password-change": {
      "post": {
        "operationId": "UsersController_changePassword",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Function"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": ["Users"]
      }
    },
    "/api/user/data-export": {
      "post": {
        "operationId": "UsersController_requestDataExport",
        "parameters": [],
        "responses": {
          "202": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": ["Users"]
      }
    },
    "/api/user/delete-request": {
      "post": {
        "operationId": "UsersController_requestDelete",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Function"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": ["Users"]
      }
    },
...
  },
  "info": {
    "title": "Psico Platform API",
    "description": "REST API for Psico Platform (psychoeducation SaaS). Design source of truth: docs/design/handoff/. Implementation plan: IMPLEMENTATION_PLAN_v2.md.",
    "version": "0.x-alpha",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "JWT access token returned by /api/auth/login"
      }
    },
    "schemas": {
...
      "Function": {
        "type": "object",
        "properties": {}
      }
    }
  }
}